Refactor Homepage tests
Rewrite test setup for new method of using hooks
Move Authform-specific tests to Authform test file
